Adding travis build file
Also moving main to it's own file so we can test linking against our own
library
