+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 2.8)
+cmake_minimum_required(VERSION 2.8.12)
 
 project(bamdb)
 
@@ -7,7 +7,7 @@ set(CMAKE_C_FLAGS "-Wall -g -std=gnu99 -fPIC")
 option(STATIC_BAMDB_BUILD "STATIC_BAMDB_BUILD" OFF)
 
 # External dependencies
-list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_LIST_DIR}/cmake")
+list(APPEND CMAKE_MODULE_PATH "${PROJECT_SOURCE_DIR}/cmake")
 find_package(CK REQUIRED)
 find_package(HTSlib REQUIRED)
 find_package(LMDB REQUIRED)
@@ -20,7 +20,7 @@ set(LIBS
   pthread
 )
 
-include_directories(include)
+include_directories("${PROJECT_SOURCE_DIR}/include")
 file(GLOB SOURCES "src/*.c")
 
 # Include dependency headers
@@ -29,10 +29,6 @@ include_directories(
   ${HTSlib_INCLUDE_DIRS}
   ${LMDB_INCLUDE_DIRS}
 )
-
-# Create bamdb executable
-add_executable(bamdb ${SOURCES})
-target_link_libraries(bamdb ${LIBS})
 
 # Create bamdb library
 if(STATIC_BAMDB_BUILD)
@@ -43,6 +39,10 @@ endif(STATIC_BAMDB_BUILD)
 
 target_link_libraries(libbamdb ${LIBS})
 set_target_properties(libbamdb PROPERTIES OUTPUT_NAME "bamdb")
+
+# Create bamdb executable
+add_executable(bamdb ${SOURCES})
+target_link_libraries(bamdb ${LIBS} libbamdb)
 
 # Default install path. Can be overridden with DESTDIR
 install(TARGETS bamdb RUNTIME DESTINATION bin)
