+++ include/bam_lmdb.h
@@ -1,21 +1,17 @@
 #ifndef BAMLMDB_H
 #define BAMLMDB_H
 
-#include <lmdb.h>
 #include <stdbool.h>
 
+#include <lmdb.h>
+
 #include "bam_api.h"
+#include "bamdb.h"
 
 /* HTSLib */
 #include "bgzf.h"
 #include "hts.h"
 #include "sam.h"
-
-typedef struct indices {
-  bool includes_qname;
-  size_t num_key_indices;  // Does not include qname index
-  char **key_indices;
-} indices_t;
 
 char *get_default_dbname(const char *filename);
